requirements.txt
pytest
pytest-catchlog
ply
amara3-iri
